refactor layerbuilder
